Filter voice list to only include voices for the current language
* Some voices match the current language *and* region - show all of them
* Some voices match the current language, but are from a different region. Add them if we
don't already have that voice. For example, in english, there might be voices with accents
from Australia or India. The user might want to use one of those.
* Some voices are provided for multiple languages & regions. I don't think it's worth
exposing that complexity in the UI.
* Some voices have the same name, but differing quality levels. Shows each quality level
available for each voice.
* Don't show voices that don't match the current language.

Looks like this app doesn't have any localization done (yet?). So this is driven off the
system language & region. I think that is the best choice, but there might be room to
improve! Especially if we knew the LLM was using a different language.