Use system default voice, unless user overrides it in Settings with explicit choice
A better default than the first voice is the voice currently chosen in System Settings.
As soon as the user chooses & saves a different voice, the app will stick with that one.

Unless that voice is uninstalled. Then the in-app Settings display picker gets out of
order (showing nothing selected and logging an error). However, the SpeechSynthesizer
class will correctly use whatever the currently selected system voice is.